+++ CHANGELOG.md
@@ -35,6 +35,7 @@
  - Added tags to openapi definition (#1751)
  - Added support to pause an agent (#1128, #1982)
  - Plugins are now imported in the inmanta_plugins package to allow importing submodules (#507)
+ - Added event listener to Environment Service (#1996)
 
 # v 2020.1 (2020-02-19) Changes in this release: